% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anti_t}
\alias{anti_t}
\title{Compute the anti transpose of a matrix (transpose with respect to the off-diagonal)}
\usage{
anti_t(m)
}
\arguments{
\item{m}{square matrix to compute the anti transpose}
}
\value{
The anti-transpose of m
}
\description{
Compute the anti transpose of a matrix (transpose with respect to the off-diagonal)
}
