% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{directUnifSampling}
\alias{directUnifSampling}
\title{Direct sampling using iid coefs}
\usage{
directUnifSampling(dag, h = 100, eps = 0.001)
}
\arguments{
\item{dag}{Graph for sampling}

\item{h}{Heat-in phase}

\item{eps}{Perturbation variance}
}
\description{
Direct sampling using iid coefs
}
