% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{.rcoef_iid}
\alias{.rcoef_iid}
\title{Sample a coefficient matrix of iid coefficients.}
\usage{
.rcoef_iid(p = 100, rmcoef = rnorm, edges)
}
\arguments{
\item{p}{Positive integer, Number of dimension.}

\item{rmcoef}{function, the random number generator for the non-zero entries.}

\item{edges}{Matrix containing the edges of the Gaussian Bayesian network (from, to)}
}
\value{
mcoef a lower triangular matrix with ones on the diagonal, and entry
(i, j) = 0 if (j, i) is not a row in \code{edges}.
}
\description{
Sample a coefficient matrix of iid coefficients.
}
\details{
Sample a sparse coefficient matrix for a Gaussian Bayesian network.
}
