% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kcontrol}
\alias{kcontrol}
\title{Control the condition number of a sample.}
\usage{
kcontrol(sam, k = 1)
}
\arguments{
\item{sam}{matrix sample already generated}

\item{k}{real number greater than \code{1}, the desired condition
number of the resulting matrix}
}
\value{
A three-dimensional array of length \code{p*p*N} with adjusted condition
number
}
\description{
Adjust the matrix sample provided so that all its elements have the desired
condition number
}
\examples{

# Generate a random undirected graph structure
ug <- igraph::sample_gnp(n = 3, p = 0.25)

# Generate 10 matrices complying with such random structure via 
# diagonal dominance 
sample <- gmat::diagdom(N = 10, ug = ug)

# Adjust condition number, for example, to 5
sample <- kcontrol(sam = sample, k = 5)

}
