% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{mh_full}
\alias{mh_full}
\title{Full Metropolis-Hastings algorithm for correlation matrices/saturated
Gaussian Bayesian networks}
\usage{
mh_full(N = 1, p = 10, h = 100, eps = 0.1)
}
\arguments{
\item{N}{Number of samples}

\item{p}{Number of variables in the saturated model}

\item{h}{Burn-in phase}

\item{eps}{Perturbation variance}
}
\description{
Full Metropolis-Hastings algorithm for correlation matrices/saturated
Gaussian Bayesian networks
}
