% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{mh_full_dag}
\alias{mh_full_dag}
\title{Full Metropolis-Hastings algorithm for Gaussian Bayesian networks}
\usage{
mh_full_dag(N = 1, dag, h = 100, eps = 0.1)
}
\arguments{
\item{N}{Number of samples}

\item{dag}{Acyclic digraph of the GBN to sample}

\item{h}{Burn-in phase}

\item{eps}{Perturbation variance}
}
\description{
Full Metropolis-Hastings algorithm for Gaussian Bayesian networks
}
