% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{mh_row}
\alias{mh_row}
\title{sampling on sphere proportionally to a power of the first coordinate}
\usage{
mh_row(N = 1, p, i = 1, h = 100, eps = 0.01, returnAll = FALSE)
}
\arguments{
\item{N}{sample size}

\item{p}{dimension}

\item{i}{exponent of the density}

\item{h}{heating phase size

Metropolis-Hasting algorithm to sample in the n dimensional semi-sphere (x_1>0)}

\item{eps}{Perturbation variance}

\item{returnAll}{Include in the output samples from the heat-in phase}
}
\description{
sampling on sphere proportionally to a power of the first coordinate
}
