% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_dag.R
\name{mh_sphere}
\alias{mh_sphere}
\title{Sampling on sphere proportionally to a power of the first coordinate

Metropolis-Hasting algorithm to sample in the
\code{p}-dimensional hemisphere.}
\usage{
mh_sphere(N = 1, p, i = 1, h = 100, eps = 0.01)
}
\arguments{
\item{N}{Number of samples.}

\item{p}{Dimension of the hemisphere.}

\item{i}{Integer, power of the first coordinate in the density.}

\item{h}{Heating phase size}

\item{eps}{Perturbation variance.}
}
\description{
Sampling on sphere proportionally to a power of the first coordinate

Metropolis-Hasting algorithm to sample in the
\code{p}-dimensional hemisphere.
}
\details{
The details of the algorithm can be found in the paper Córdoba et al. (2018), including a discussion on theoretical convergence and numerical experiments for choosing its hyper parameters.
}
\examples{
mh_sphere(N = 4, p = 3, i = 2)
}
\references{
Córdoba I., Varando G., Bielza C., Larrañaga P. A fast
Metropolis-Hastings method for generating random correlation matrices. \emph{Lecture Notes in
Computer Science} (IDEAL 2018), vol 11314, pp. 117-124, 2018.
}
\author{
Gherardo Varando \email{gherardo.varando@math.ku.dk}
}
