% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{.rcoef_polar}
\alias{.rcoef_polar}
\alias{.sin_k}
\alias{.sin_k_cum}
\alias{.sin_int}
\alias{.rsin}
\title{Sample coefficient matrix with polar parametrization}
\usage{
.rcoef_polar(p = 100, method = "numeric", L)

.sin_k(x, k)

.sin_k_cum(x, k, method = "numeric")

.sin_int(x, k = 1)

.rsin(n, k = 1, method = "numeric")
}
\arguments{
\item{p}{positive integer, the dimension of the square matrix}

\item{method}{String, method to sample from \code{sin(x)^k}: \code{numeric} or
\code{recursive}}

\item{x}{value where \code{sin^k} will be calculated}

\item{k}{exponent for \code{sin^k}}

\item{n}{Number of samples to generate}

\item{edges}{Matrix defining the edges of the Bayesian network (from, to).}
}
\value{
\code{mcoef} a lower triangular matrix such that mcoefmcoef' has 1 on diagonal.
}
\description{
Sample a polar parametrization of the Cholesky factor in the LL' decomposition of
the concentration matrix.
}
\details{
The mcoef factor is sampled such that mcoefmcoef' is uniformly
distributed among the correlation matrices.

recursive computation
}
