% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{rgbn}
\alias{rgbn}
\title{Sample a random Gaussian Bayesian network}
\usage{
rgbn(p = 10, method = "iidcoef", d = 0.25, rmcoef = rnorm,
  comp = "numeric", dag = NULL)
}
\arguments{
\item{p}{Positive integer, Number of dimension.}

\item{method}{String one of "iidcoef", "unifconc"}

\item{d}{number in \code{[0,1]}, the proportion of non-zero entries.}

\item{rmcoef}{function, the random number generator for the non-zero
entries of the coefficient matrix for the "iidcoef" method.}

\item{comp}{String one of "numeric" or "recursive", indicating the
computational method to use for sampling the angles for "unifconc" method}

\item{dag}{A graph object of class \code{igraph}.}
}
\value{
A list with the following components:
\itemize{
\item param A list with the parameters for the concentration matrix. For
"iidcoef", these are the coefficient matrix and the conditional variance
matrix. For "unifconc", the parameters are the angle matrix and the
coefficient matrix.
\item mconc The sampled concentration matrix
\item dag The resulting dag structure
}
}
\description{
Samples a random Gaussian Bayesian network with different methods.
}
