% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{rgbn_iid}
\alias{rgbn_iid}
\title{Sample a random Gaussian Bayesian network}
\usage{
rgbn_iid(N = 1, p = 10, dag = NULL, return.minvector = FALSE)
}
\arguments{
\item{N}{Number of samples}

\item{p}{Positive integer, Number of dimension.}

\item{dag}{A graph object of class \code{igraph}.}

\item{return.minvector}{logical, if TRUE the minimimal vector representation is returned (useful to plot in the elliptope)}
}
\value{
The sample
}
\description{
Samples a random Gaussian Bayesian network with iid coefficients
}
