% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{rgbn_polar}
\alias{rgbn_polar}
\title{Sample a random Gaussian Bayesian network}
\usage{
rgbn_polar(N = 1, p = 10, comp = "numeric", dag = NULL,
  return.minvector = FALSE)
}
\arguments{
\item{N}{Number of samples}

\item{p}{Positive integer, Number of dimension.}

\item{comp}{String one of "numeric" or "recursive", indicating the
computational method to use for sampling the angles for "unifconc" method}

\item{dag}{A graph object of class \code{igraph}.}

\item{return.minvector}{logical, if TRUE the minimimal vector representation is returned (useful to plot in the elliptope)}
}
\value{
The sample
}
\description{
Samples a random Gaussian Bayesian network with polar parametrization.
}
