% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmn_sampling.R
\name{rgmn}
\alias{rgmn}
\title{Sample a random Gaussian Markov network}
\usage{
rgmn(N = 1, ug = NULL, p = 10, d = 0.25, method = "sqrt",
  rentries = runif, k = NULL, zapzeros = FALSE)
}
\arguments{
\item{N}{sample size.}

\item{ug}{the undirected graph}

\item{p}{Positive integer, Number of dimension.}

\item{d}{number in \code{[0,1]}, the proportion of non-zero
entries.}

\item{method}{string one of the following: \code{sqrt}, \code{domdiag}.}

\item{rentries}{function, the random number generator for the
non-zero entries off the diagonal.}

\item{k}{real number greater than \code{1}, the desired condition
number of the resulting matrix.}

\item{zapzeros}{convert to zero extremely low entries.}
}
\value{
A list with the following components:
\itemize{
\item \code{param} A list with the parameters for the concentration
matrix. In both cases, it consists on a square matrix and
a diagonal matrix
\item \code{mconc} The sampled concentration matrix.
}
}
\description{
Samples the concentration matrix corresponding to a Gaussian
Markov network by two different methods: adding a strictly
positive diagonal to the square root of a matrix, or adding
a dominant diagonal to an arbitrary square matrix.
}
\details{
Sample a sparse symmetric positive definite matrix.
}
