% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{rUnifChordalConc}
\alias{rUnifChordalConc}
\title{rUnifChordalConc}
\usage{
rUnifChordalConc(N = 1, dag = NULL, return.minvector = FALSE, ...)
}
\arguments{
\item{N}{Number of samples}

\item{dag}{directed chordal acyclic graph, use the igraph package graph class or alternatively
a matrix that will be interpreted as an adjacency matrix.}

\item{return.minvector}{logical, if TRUE the minimimal vector representation is returned (useful to plot in the elliptope)}

\item{...}{additional parameters}
}
\description{
rUnifChordalConc
}
