% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_cond_number}
\alias{set_cond_number}
\title{Set the condition number of the matrices in a sample of covariance/correlation matrices}
\usage{
set_cond_number(sample, k)
}
\arguments{
\item{sample}{Array, the \verb{p x p x N} matrix sample.}

\item{k}{Condition number to be set.}
}
\value{
A \verb{p x p x N} array containing the matrices with the fixed condition
number.
}
\description{
Set the condition number of the matrices in a sample of covariance/correlation matrices
}
