% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{uchol}
\alias{uchol}
\title{uchol}
\usage{
uchol(N = 1, p = 10, return.minvector = FALSE, ...)
}
\arguments{
\item{N}{}

\item{return.minvector}{logical, if TRUE the minimimal vector representation
is returned (useful to plot in the elliptope)}

\item{...}{additional parameters}
}
\description{
uchol
}
