% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ug_to_dag}
\alias{ug_to_dag}
\title{Minimal DAG from UG}
\usage{
ug_to_dag(ug)
}
\arguments{
\item{x}{igraph graph or adjacency matrix}
}
\value{
acyclic directed graph orientation (igraph)
}
\description{
Find the DAG with same skeleton as the UG and
with the minimum number of v-structures added.
}
