% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ugTwodag}
\alias{ugTwodag}
\title{Minimal DAG from UG}
\usage{
ugTwodag(x)
}
\arguments{
\item{x}{igraph graph or adjacency matrix}
}
\value{
adjacency matrix for directed graph
}
\description{
find the DAG with same skeleton as the UG and
with the minimum number of v-structures added.
}
