% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{directed graph matrix simulation}
\alias{directed graph matrix simulation}
\alias{chol_mh}
\alias{chol_iid}
\alias{chol_polar}
\title{Sample a random Gaussian Bayesian network using different parametrizations
for the Cholesky decomposition.}
\usage{
chol_mh(N = 1, p = 10, dag = NULL, return.minvector = FALSE,
  add_no_chordal = TRUE, ...)

chol_iid(N = 1, p = 10, dag = NULL, return.minvector = FALSE)

chol_polar(N = 1, p = 10, comp = "numeric", dag = NULL,
  return.minvector = FALSE)
}
\arguments{
\item{N}{Number of samples}

\item{p}{Number of variables}

\item{dag}{directed chordal acyclic graph, use the igraph package graph class}

\item{return.minvector}{logical, if TRUE the minimimal vector representation
is returned (useful to plot in the elliptope)}

\item{add_no_chordal}{logical, if TRUE when the dag provided is not chordal,
a fill-in is computed, in order to ensure uniform distribution}

\item{...}{additional parameters}

\item{comp}{String one of "numeric" or "recursive", indicating the
computational method to use for sampling the angles for "unifconc" method}
}
\value{
A three-dimensional array of length \code{p*p*N}
}
\description{
Sample a random Gaussian Bayesian network using different parametrizations
for the Cholesky decomposition.
}
\details{
Function \code{rgbn_chol} uses the method described in Córdoba et al.
(2018), based on a Metropolis-Hastings algorithm over the upper Cholesky
factorization.

The entries in the upper Cholesky factor are sampled i.i.d. by
function \code{rgbn_iid}, following Kalisch and Buhlmann (2007).

Function \code{rgbn_polar} reparametrizes the Cholesky factor following
the approach by Pourahmadi and Wang (2015).
}
\examples{

# Generate a random acyclic digraph structure
dag <- gmat::rgraph(p = 3, d = 0.25, dag = TRUE)

# Generate 2 matrices via Cholesky decomposition
chol_mh(N = 2, dag = dag, add_no_chordal = FALSE)

# Generate 2 matrices with i.i.d. entries in the Cholesky factor
chol_iid(N = 2, dag = dag)

# Generate 2 matrices using the polar parametrization of the Cholesky factor
chol_polar(N = 2, dag = dag)

}
\references{
Córdoba I., Varando G., Bielza C., Larrañaga P. A fast
Metropolis-Hastings method for generating random correlation matrices. Intelligent Data
Engineering and Automated Learning – IDEAL 2018. Lecture Notes in
Computer Science, vol 11314, pp. 117-124, 2018.

Kalisch, M., Buhlmann, P. Estimating high-dimensional directed
acyclic graphs with the PC-algorithm, Journal of Machine Learning Research,
8:613-636, 2007.

Pourahmadi, M., Wang, X. Distribution of random correlation matrices:
Hyperspherical parameterization of the Cholesky factor, Statistics &
Probability Letters, 106:5-12, 2015.
}
