% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rgmat}
\alias{rgmat}
\alias{port}
\alias{diagdom}
\title{Sample random covariance/concentration graph matrices.}
\usage{
port(N = 1, ug = NULL, p = 5, d = 0.25, rentries = runif,
  zapzeros = TRUE)

diagdom(N = 1, ug = NULL, p = 5, d = 0.25, rentries = runif)
}
\arguments{
\item{N}{sample size}

\item{ug}{the undirected graph}

\item{p}{Matrix dimension (number of nodes in \code{ug} if provided)}

\item{d}{number in \code{[0,1]}, the proportion of non-zero
entries (if \code{ug} is not provided)}

\item{rentries}{the random number generator for the
non-zero entries (defaults to \code{runif})}

\item{zapzeros}{convert to zero extremely low entries}
}
\value{
A three-dimensional array of length \code{p*p*N}
}
\description{
Samples the concentration/covariance matrix corresponding to an
undirected graph by either partial orthogonalization or diagonal dominance.
}
