% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmn_sampling.R
\name{rgmn}
\alias{rgmn}
\title{Sample a random Gaussian Markov network}
\usage{
rgmn(N = 1, ug = NULL, p = 10, d = 0.25, method = "port",
  rentries = runif, k = NULL, zapzeros = FALSE)
}
\arguments{
\item{N}{sample size}

\item{ug}{the undirected graph}

\item{p}{Matrix dimension (number of nodes in \code{ug} if provided)}

\item{d}{number in \code{[0,1]}, the proportion of non-zero
entries (if \code{ug} is not provided)}

\item{method}{one of the following: \code{port} (partial orthogonalization),
\code{diagdom} (diagonal dominance)}

\item{rentries}{the random number generator for the
non-zero entries (defaults to \code{runif})}

\item{k}{real number greater than \code{1}, the desired condition
number of the resulting matrix}

\item{zapzeros}{convert to zero extremely low entries}
}
\value{
A three-dimensional array of length \code{p*p*N}
}
\description{
Samples the concentration/covariance matrix corresponding to an
undirected graph by two different methods: partial orthogonalization
or diagonal dominance.
}
