% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{sphereSample}
\alias{sphereSample}
\title{sampling on sphere proportionally to a power of the first coordinate}
\usage{
sphereSample(N = 1, n, k = 1, h = 100, s = 1, eps = 0.01,
  returnAll = FALSE)
}
\arguments{
\item{N}{sample size}

\item{n}{dimension}

\item{k}{exponent of the density}

\item{h}{heating phase size}

\item{s}{step

Metropolis-Hasting algorithm to sample in the n dimensional semi-sphere (x_1>0)}

\item{eps}{Perturbation variance}

\item{returnAll}{Include in the output samples from the heat-in phase}
}
\description{
sampling on sphere proportionally to a power of the first coordinate
}
